+++ tests/sdk/v2/api/__init__.py
@@ -1,0 +1,6 @@
+################################################################################
+# © Copyright 2023 Zapata Computing Inc.
+################################################################################
+"""
+Required by pytest for relative imports.
+"""
